bugs fixes

+++ src/pages/StorePackDetails.js
@@ -1,36 +1,37 @@
 import React, { useContext, useMemo, useState, useEffect } from 'react'
-import { Block, Page, Navbar, Card, CardContent, Icon, Fab, Toolbar, FabButtons, FabButton, CardFooter} from 'framework7-react'
-import BottomToolbar from './BottomToolbar'
+import { Block, Page, Navbar, Card, CardContent, Toolbar, CardFooter, Popover, List, Link, ListItem } from 'framework7-react'
 import { StoreContext } from '../data/Store';
-import { deleteStorePack, confirmPrice, haltOffer, showMessage, showError, getMessage } from '../data/Actions'
+import { deleteStorePack, confirmPrice, haltOffer, extendOffer, showMessage, showError, getMessage } from '../data/Actions'
 
 const StorePackDetails = props => {
   const { state } = useContext(StoreContext)
   const [error, setError] = useState('')
   const storePack = useMemo(() => state.storePacks.find(p => p.id === props.id)
   , [state.storePacks, props.id])
-  const pack = useMemo(() => state.packs.find(p => p.id === storePack.packId)
+  const pack = useMemo(() => storePack ? state.packs.find(p => p.id === storePack.packId) : ''
   , [state.packs, storePack])
-  const product = useMemo(() => state.products.find(p => p.id === pack.productId)
+  const product = useMemo(() => pack ? state.products.find(p => p.id === pack.productId) : ''
   , [state.products, pack])
+
   useEffect(() => {
     if (error) {
       showError(props, error)
       setError('')
     }
   }, [error, props])
-
   const handleEditPrice = () => {
     props.f7router.navigate(`/editPrice/${storePack.id}`)
   }
-  const handleDelete = async () => {
-    try{
-      await deleteStorePack(storePack, pack, state.storePacks)
-      showMessage(props, state.labels.deleteSuccess)
-      props.f7router.back()
-    } catch(err) {
-			setError(getMessage(err, state.labels, props.f7route.route.component.name))
-		}
+  const handleDelete = () => {
+    props.f7router.app.dialog.confirm(state.labels.confirmationText, state.labels.confirmationTitle, async () => {
+      try{
+        await deleteStorePack(storePack, pack, state.storePacks)
+        showMessage(props, state.labels.deleteSuccess)
+        props.f7router.navigate('/home/', {reloadAll: true})
+      } catch(err) {
+        setError(getMessage(err, state.labels, props.f7route.route.component.name))
+      }
+    })
   }
   const handleConfirmPrice = async () => {
     try{
@@ -45,23 +46,46 @@ const StorePackDetails = props => {
     try{
       const offerEndDate = new Date(storePack.offerEnd)
       const today = (new Date()).setHours(0, 0, 0, 0)
-      let confirmation
       if (offerEndDate > today) {
-        props.f7router.app.dialog.confirm(state.labels.confirmationText, () => {
-          confirmation = true
+        props.f7router.app.dialog.confirm(state.labels.confirmationText, state.labels.confirmationTitle, async () => {
+          try{
+            await haltOffer(storePack, pack, state.storePacks)
+            showMessage(props, state.labels.haltSuccess)
+            props.f7router.back()  
+          } catch(err) {
+            setError(getMessage(err, state.labels, props.f7route.route.component.name))
+          }
         })
       } else {
-        confirmation = true
+        await haltOffer(storePack, pack, state.storePacks)
+        showMessage(props, state.labels.haltSuccess)
+        props.f7router.back()    
       }
-      if (!confirmation) return
-      await haltOffer(storePack, pack, state.storePacks)
-      showMessage(props, state.labels.haltSuccess)
-      props.f7router.back()  
     } catch(err) {
 			setError(getMessage(err, state.labels, props.f7route.route.component.name))
 		}
   }
-
+  const handleExtendOffer = () => {
+    props.f7router.app.dialog.prompt(state.labels.confirmationText, state.labels.confirmationTitle, async days => {
+      try{
+        if (!days || Number(days) <= 0) {
+          throw new Error('invalidValue')
+        }
+        const offerEnd = storePack.offerEnd.toDate()
+        offerEnd.setDate(offerEnd.getDate() + Number(days))
+        const newStorePack = {
+          ...storePack,
+          offerEnd
+        }
+        await extendOffer(newStorePack)
+        showMessage(props, state.labels.editSuccess)
+        props.f7router.back() 
+      } catch(err) {
+        setError(getMessage(err, state.labels, props.f7route.route.component.name))
+      }
+    })
+  }
+  if (!storePack) return null //to handle delete operation
   return (
     <Page>
       <Navbar title={`${product.name} - ${state.stores.find(s => s.id === storePack.storeId).name}`} backLink={state.labels.back} />
@@ -76,29 +100,19 @@ const StorePackDetails = props => {
           </CardFooter>
         </Card>
       </Block>
-      <Fab position="left-top" slot="fixed" color="orange">
-        <Icon material="keyboard_arrow_down"></Icon>
-        <Icon material="close"></Icon>
-        <FabButtons position="bottom">
-          <FabButton color="blue" onClick={() => handleEditPrice()}>
-            <Icon material="edit"></Icon>
-          </FabButton>
-          <FabButton color="green" onClick={() => handleConfirmPrice()}>
-            <Icon material="done"></Icon>
-          </FabButton>
-          <FabButton color="red" onClick={() => handleDelete()}>
-            <Icon material="delete"></Icon>
-          </FabButton>
-          {storePack.offerEnd ? 
-            <FabButton color="red" onClick={() => handleHaltOffer()}>
-              <Icon material="report_problem"></Icon>
-            </FabButton>
-            : ''
-          }
-        </FabButtons>
-      </Fab>
+      <Popover className="popover-menu">
+        <List>
+          <ListItem link="#" popoverClose title={state.labels.editPrice} onClick={() => handleEditPrice()}/>
+          <ListItem link="#" popoverClose title={state.labels.delete} onClick={() => handleDelete()}/>
+          <ListItem link="#" popoverClose title={state.labels.confirmPrice} onClick={() => handleConfirmPrice()}/>
+          {storePack.offerEnd ? <ListItem link="#" popoverClose title={state.labels.haltOffer} onClick={() => handleHaltOffer()}/> : ''}
+          {storePack.offerEnd ? <ListItem link="#" popoverClose title={state.labels.extendOffer} onClick={() => handleExtendOffer()}/> : ''}
+        </List>
+      </Popover>
+
       <Toolbar bottom>
-        <BottomToolbar/>
+        <Link href="/home/" iconMaterial="home" />
+        <Link popoverOpen=".popover-menu" iconMaterial="more_vert" />
       </Toolbar>
     </Page>
   )
